+++ README_Testen.md
@@ -1,0 +1,93 @@
+### 1. **Unit Tests (Einheitstests)**
+   - **Laravel (Backend):**
+     - **Modelle**: Teste die Modelle, um sicherzustellen, dass sie korrekt mit der Datenbank interagieren. Dies umfasst das Testen von Beziehungen, Mutatoren und Zugriffsmethoden.
+     - **Services**: Teste individuelle Service-Klassen und deren Methoden.
+     - **Helper-Funktionen**: Teste alle benutzerdefinierten Helper-Funktionen, die du geschrieben hast.
+
+   - **React (Frontend):**
+     - **Komponenten**: Teste einzelne React-Komponenten. Stelle sicher, dass sie korrekt rendern, insbesondere mit verschiedenen Props.
+     - **Utility-Funktionen**: Teste alle Funktionen, die in React-Komponenten verwendet werden, um sicherzustellen, dass sie das gewünschte Verhalten zeigen.
+
+   **Tools:**
+   - Laravel: PHPUnit (integriert in Laravel)
+   - React: Jest, React Testing Library
+
+### 2. **Feature Tests (Funktions- und Integrations-Tests)**
+   - **Laravel:**
+     - **Controller**: Teste die Routen und die Controller-Logik, um sicherzustellen, dass Endpunkte korrekt arbeiten.
+     - **Middleware**: Teste Middleware, um sicherzustellen, dass sie korrekt ausgelöst wird (z. B. Authentifizierung, Validierung).
+     - **Datenbank**: Teste die Integration mit der Datenbank, einschließlich Abfragen, Einfügeoperationen und Aktualisierungen.
+
+   - **React:**
+     - **Komponenten-Interaktion**: Teste das Zusammenspiel verschiedener Komponenten, insbesondere bei komplexen Formulierungen und Benutzerinteraktionen.
+     - **Routing**: Teste den React-Router und die Navigation innerhalb der Anwendung.
+
+   **Tools:**
+   - Laravel: PHPUnit, Laravel Dusk (für Browser-Tests)
+   - React: Jest, React Testing Library
+
+### 3. **End-to-End (E2E) Tests**
+   - Diese Art von Tests überprüft den gesamten Ablauf der Anwendung von Anfang bis Ende, einschließlich des Frontends und des Backends.
+   - **Beispiel-Szenarien**: Benutzeranmeldung, Dateneingabeformulare, vollständige Workflows, bei denen Benutzerinteraktionen das Backend und das Frontend durchlaufen.
+
+   **Tools:**
+   - Cypress, Selenium, Laravel Dusk (speziell für Laravel)
+
+### 4. **Datenbanktests**
+   - Teste Migrationen, Seeder und Datenintegrität.
+   - Überprüfe, ob alle Beziehungen (z. B. 1:n, n:m) korrekt funktionieren.
+   - Verwende Testdatenbanken oder SQLite in-memory-Datenbanken für Tests.
+
+   **Tools:**
+   - Laravel: PHPUnit, Datenbank-Factories
+
+### 5. **API-Tests**
+   - Teste RESTful APIs, um sicherzustellen, dass sie korrekt funktionieren.
+   - Überprüfe HTTP-Statuscodes, Antwortzeiten und Antwortstrukturen.
+   - Führe Tests für verschiedene Rollen oder Benutzerrechte durch.
+
+   **Tools:**
+   - Postman, PHPUnit (mit `assertJson` für JSON-Antworten)
+
+### 6. **Performance-Tests**
+   - Teste die Leistung der Anwendung unter Last.
+   - Simuliere mehrere gleichzeitige Benutzer und überprüfe die Reaktionszeit und das Verhalten der Anwendung.
+
+   **Tools:**
+   - Apache JMeter, Laravel Telescope, k6
+
+### 7. **Sicherheitstests**
+   - Überprüfe auf häufige Sicherheitslücken wie SQL-Injection, XSS, CSRF.
+   - Überprüfe den Authentifizierungs- und Autorisierungsprozess.
+   - Stelle sicher, dass sensible Daten verschlüsselt gespeichert werden.
+
+   **Tools:**
+   - OWASP ZAP, PHPUnit (für Sicherheits- und Authentifizierungstests)
+
+### 8. **Usability-Tests (Benutzerfreundlichkeitstests)**
+   - Teste das UI und UX der Anwendung, um sicherzustellen, dass es benutzerfreundlich und intuitiv ist.
+   - Führe Tests mit echten Benutzern durch, um ihre Erfahrungen zu messen und Verbesserungen vorzunehmen.
+
+   **Tools:**
+   - Manuelle Tests, Benutzerfeedback
+
+### 9. **Browser-Kompatibilitätstests**
+   - Stelle sicher, dass die Anwendung in verschiedenen Browsern und auf verschiedenen Geräten korrekt funktioniert.
+   - Überprüfe insbesondere auf mobile Responsivität.
+
+   **Tools:**
+   - BrowserStack, Sauce Labs
+
+### 10. **Automatisierte Tests (Continuous Integration/Continuous Deployment - CI/CD)**
+   - Richte ein CI/CD-Pipeline ein, um Tests bei jedem Code-Push automatisch auszuführen.
+   - Sicherstellen, dass keine Änderungen in den Haupt-Branch gelangen, ohne dass alle Tests bestanden wurden.
+
+   **Tools:**
+   - GitHub Actions, GitLab CI, Jenkins, CircleCI
+
+### 11. **Dokumentation**
+   - Dokumentiere alle Tests, damit das Team versteht, was getestet wurde und wie die Tests durchgeführt werden.
+   - Dokumentiere auch alle Testergebnisse und bekannte Probleme.
+
+### Fazit
+Der Testprozess sollte sowohl Unit-Tests als auch Integrationstests umfassen, die eine Kombination aus automatisierten und manuellen Tests beinhalten. Verwende die oben genannten Tools und Best Practices, um sicherzustellen, dass deine Laravel- und React-Anwendung robust und fehlerfrei ist.
