--- tests/Feature/Groceries/GroceriesTest.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Tests\Feature\Orders;
-
-use Tests\TestCase;
-
-class GroceriesTest extends TestCase
-{
-    public function test_example(): void
-    {
-        $this->assertTrue(true);
-    }
-}
