--- tests/Feature/User/UserManagementTest.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Tests\Feature\User;
-
-use Tests\TestCase;
-
-class UserManagemenTest extends TestCase
-{
-    public function test_example(): void
-    {
-        $this->assertTrue(true);
-    }
-}
